requirements.txt
apache-beam[gcp]==2.50.0
google-cloud-language==2.10.1
google-cloud-pubsub==2.18.1
google-cloud-bigquery==3.11.4
